feat: Enable HTTPS support for server
This commit adds support for HTTPS in the server by implementing the necessary logic to create an HTTPS server using the provided SSL key and certificate. The server now checks the `USE_HTTPS` environment variable to determine whether to use HTTPS or HTTP. The commit also updates the server startup message to display the correct protocol (HTTP or HTTPS) based on the environment variable.